+++ CdIts.Caffoa.Cli/Generator/InterfaceGenerator.cs
@@ -85,7 +85,7 @@ public class InterfaceGenerator
 
     public List<ParameterBuilder.Overload> GetParams(EndPointModel endpoint) => GetMethodParams(endpoint, _config);
     
-    public static List<ParameterBuilder.Overload> GetMethodParams(EndPointModel endpoint, CaffoaConfig config, bool withDurable = true, bool nullableDefaults = false, bool addAspNetAttributes = false)
+    public static List<ParameterBuilder.Overload> GetMethodParams(EndPointModel endpoint, CaffoaConfig config, bool withDurable = true, bool nullableDefaults = false, bool addAspNetAttributes = false, bool forceCancellation = false)
     {
         var builder = ParameterBuilder.Instance(config.UseDateOnly is true && config.ParsePathParameters is not false, config.UseDateTime is true, addAspNetAttributes)
             .AddPathParameters(endpoint.Parameters);
@@ -95,7 +95,7 @@ public class InterfaceGenerator
             builder.AddDurableClient();
         if (config.ParseQueryParameters is not false)
             builder.AddQueryParameters(endpoint.QueryParameters(), nullableDefaults);
-        if (config.WithCancellation is not false)
+        if (config.WithCancellation is not false || forceCancellation)
             builder.AddCancellationToken();
         if (endpoint.HasRequestBody)
         {